bugs fix

+++ l2fwd/libsashstore/conshash.h
@@ -10,7 +10,10 @@
 #define TABLE_SIZE 65537
 
 typedef int Node;
-typedef int8_t LookUpTable[TABLE_SIZE];
+/* Holds a backend identifier that callers use as a MAC value. Must be wider
+ * than a byte: as int8_t, the 0xff written by populate_lut() sign-extended to
+ * -1 in every caller, so every flow mapped to the same all-ones pseudo-MAC. */
+typedef int64_t LookUpTable[TABLE_SIZE];
 
 void populate_lut(LookUpTable lut);
